Update homepage and add tags page
Update the homepage to be more inviting. Adds my Twitter, GitHub and LinkedIn
information, a gravatar image that will automatically update when Gravatar
images are changed, and cleans up the content a bit.

Tags are now available for every blog post (!!!), which might make it easier to
sort content for people who are into that sort of thing.