Correct image path for post
A post that had an image was changed to another date, and the image
wasn't moved to a similarly path'd date.

**The TL;DR is:** there is a helper in blog show posts that puts all
images for said post into a folder, `img/20xx-xx-xx-post-title`, which
is meant to keep the image folder from getting unruly. However, we need
to remember this helper when moving posts in the future!