Reconfigure blog index page
Reconfigure the blog index page to resemble Ghost's [casper
theme][casper], which is a better UX experience.

Remove search option from the site, which went largely unused. 95% of
users who visit the site currently are direct referrals searching Google
for a topic and coming across the site organically. This option can be
added in later if needed.

Add categories for all posts. This is necessary for the blog index page,
so the correct image appears in the post data. If a post has an image
specified in the Jekyll topmatter, this image will be used; otherwise,
the image will default to the category image.

[casper]: https://github.com/TryGhost/Casper