Redesign!
A full redesign featuring Tailwind CSS, a light sprinkling of VueJS and
featuring Turbolinks for faster page loads.

I've been sitting on this redesign for several months now. This morning,
after deciding to write a new blog post, I found the repo in a state of
almost-completed-but-not-commited and decided to launch.

Onward and upward!