Update site config for blog functionality
Remove some of the defaults that come with a new Gatsby install. Add
some things that will be needed for good internet citizenship: sitemaps,
actual CSS, etc.

One of the largest changes in this commit is the `gatsby-node.js` file:
it has been configured to create blog posts for every `*.md` file in the
content directory. Each directory is treated as a separate "content
type" instance, so I can have `/blog`, `/journal`, and `/til` as
separate blog entry points.