Add styles and layout to site
Remove some default styles dropped in by Gatsby, and replace with our
own styles added by Tailwind.

Add a new helper for links, `PartialNavLink`, which is used in the blog
sections to trigger an active class if the URL contains a fragment of
the URL within the `PartialNavLink` element.

Override `html.js` so we can add classes directly onto the root
container.